Move OTel basic setup stuff to utils. Add simple auto instrumentation.
