+++ pkg/util/otel/instrumentation.go
@@ -1,0 +1,1 @@
+package otel
